+++ examples/resources/iosxe_pim_vrf/resource.tf
@@ -18,4 +18,12 @@ resource "iosxe_pim_vrf" "example" {
       bidir       = false
     }
   ]
+  rp_candidates = [
+    {
+      interface = "Loopback100"
+      interval  = 100
+      priority  = 10
+      bidir     = false
+    }
+  ]
 }