+++ internal/provider/resource_iosxe_pim_vrf_test.go
@@ -27,10 +27,15 @@ func TestAccIosxePIMVRF(t *testing.T) {
 					resource.TestCheckResourceAttr("iosxe_pim_vrf.test", "rp_address", "19.19.19.19"),
 					resource.TestCheckResourceAttr("iosxe_pim_vrf.test", "rp_address_override", "false"),
 					resource.TestCheckResourceAttr("iosxe_pim_vrf.test", "rp_address_bidir", "false"),
+					resource.TestCheckResourceAttr("iosxe_pim_vrf.test", "cache_rpf_oif", "true"),
 					resource.TestCheckResourceAttr("iosxe_pim_vrf.test", "rp_addresses.0.access_list", "10"),
 					resource.TestCheckResourceAttr("iosxe_pim_vrf.test", "rp_addresses.0.rp_address", "10.10.10.10"),
 					resource.TestCheckResourceAttr("iosxe_pim_vrf.test", "rp_addresses.0.override", "false"),
 					resource.TestCheckResourceAttr("iosxe_pim_vrf.test", "rp_addresses.0.bidir", "false"),
+					resource.TestCheckResourceAttr("iosxe_pim_vrf.test", "rp_candidates.0.interface", "Loopback100"),
+					resource.TestCheckResourceAttr("iosxe_pim_vrf.test", "rp_candidates.0.interval", "100"),
+					resource.TestCheckResourceAttr("iosxe_pim_vrf.test", "rp_candidates.0.priority", "10"),
+					resource.TestCheckResourceAttr("iosxe_pim_vrf.test", "rp_candidates.0.bidir", "false"),
 				),
 			},
 			{
@@ -88,10 +93,17 @@ func testAccIosxePIMVRFConfig_all() string {
 		rp_address = "19.19.19.19"
 		rp_address_override = false
 		rp_address_bidir = false
+		cache_rpf_oif = true
 		rp_addresses = [{
 			access_list = "10"
 			rp_address = "10.10.10.10"
 			override = false
+			bidir = false
+		}]
+		rp_candidates = [{
+			interface = "Loopback100"
+			interval = 100
+			priority = 10
 			bidir = false
 		}]
 		depends_on = [iosxe_restconf.PreReq0, iosxe_restconf.PreReq1, ]